+++ serde_json_path/CHANGELOG.md
@@ -7,13 +7,21 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 # Unreleased
 
+# 0.5.2 (13 March 2023)
+
 - **added:** Add `first`, `last`, and `get` methods to `NodeList` type ([#16])
 - **changed:** Make `NodeList::at_most_one` and `NodeList::exactly_one` take `&self` instead of `self` ([#16])
 - **docs:** Update crate-level docs to better reflect recent changes ([#21])
 - **docs:** Corrected a broken link in crate-level docs ([#21])
+- **added:** derive `Clone` for `JsonPath` and its descendants ([#24])
+- **added:** derive `Default` for `JsonPath` ([#25])
+- **added:** implement `Display`  and `Serialize` for `JsonPath` ([#26])
 
 [#16]: https://github.com/hiltontj/serde_json_path/pull/16
 [#21]: https://github.com/hiltontj/serde_json_path/pull/21
+[#24]: https://github.com/hiltontj/serde_json_path/pull/24
+[#25]: https://github.com/hiltontj/serde_json_path/pull/25
+[#26]: https://github.com/hiltontj/serde_json_path/pull/26
 
 # 0.5.1 (11 March 2023)